VIPBOT/project_zaid/bot/requirements.txt
flask
flask_restful
aiohttp
aiofiles
bs4
googlesearch-python==1.0.1
gitpython
humanize
httpx
heroku3
requests
motor==3.0.0
opencv-python-headless
pyrogram==2.0.38
pymongo==4.3.2
python-dotenv
pymediainfo
prettytable
git+https://github.com/Itz-fork/py-trans.git
reportlab
svglib
speedtest-cli==2.1.3
tgcrypto 
telegraph
uwuify
pyaes==1.6.1
pysocks==1.7.1
aiosqlite>=0.16.0,<0.18.0
python-dotenv
